Use bracketed paste mode when sending to replxx
This change makes the Execute Selection/Line commands send text to the
terminal wrapped in bracketed paste [1] control characters, if the bqn
executable is built with replxx (assumed true by default, configurable).
This means it behaves the same as copying and pasting into the REPL: it
treats the input as a whole, rather than trying to execute each line.
For example, if you select these three lines and hit Ctrl+Enter:

    F ← {
      𝕨+𝕩
    }
    1 F 2

Before this change, it would result in:

    "F ← {" Error: Unmatched bracket
    "  𝕨+𝕩" Error: Special name outside of any block
    "}"     Error: Unmatched bracket
    "1 F 2" Error: Undefined identifier

But now it results in 3 as expected.

This also enables getting rid of the "bqn.sendLoadScriptNewlineDelay"
hack. "bqn.sendToNewReplDelay" still seems to be necessary.

In the future the sendText API for vscode.Terminal will support a third
parameter, bracketedPasteMode [2]. When it does, we should switch to
that instead of manually adding escape codes.

[1] https://en.wikipedia.org/wiki/Bracketed-paste
[2] https://github.com/microsoft/vscode/issues/153592#issuecomment-1182045382