Handle surrogate pairs in hover provider (Fixes #19)
This commit make the hover provider always check the full Unicode code
point under the cursor. Before, it only looked at one UTF-16 code unit,
which didn't work for characters such as "𝕩" which are encoded in UTF-16
as a surrogate pair.

This also makes the hover provider work for both cases of these
characters, e.g. "𝕏" instead of "𝕩", which is documented on the same
help page as "𝕩" rather than getting its own page.