Show documentation for system values on hover
I extended gen_help.sh to parse documentation from BQN/spec/system.md.
For docs not formatted in tables I had to write them manually.

If you hover on the bullet, you see the bullet docs, but if you hover
anywhere on the word itself, then you see the system value docs.

A future improvement could be to also add these to snippets so that they
autocomplete as well.

Fixes #30